requirements.txt
apache-airflow==2.6.3
requests
psycopg2-binary